Changed float to int for MaxClients() as part of vscript as it makes more sense for this function

+++ src/game/server/vscript_server.cpp
@@ -1006,7 +1006,7 @@ static float Time()
 	return gpGlobals->curtime;
 }
 
-static float MaxClients()
+static int MaxClients()
 {
 	return gpGlobals->maxClients;
 }
